requirements.txt
Flask
pandas
scikit-learn
yfinance
spacy
requests
python-dotenv
shap
vaderSentiment